Time: 178 ms (86.92%), Space: 92.5 MB (14.87%) - LeetHub

+++ 2344-minimum-deletions-to-make-array-divisible/2344-minimum-deletions-to-make-array-divisible.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minOperations(vector<int>& nums, vector<int>& numsDivide) {
-        int g = numsDivide[0], ans =0;
+        size_t g = numsDivide[0], ans =0;
         for(int i=1;i < numsDivide.size();++i)
         g = gcd(g,numsDivide[i]);
-        map<int,int> m;
+        map<size_t,size_t> m;
         for(auto &it : nums) m[it]++;
         for(auto &it: m)
         {
